### 2019-01-22 #1196 @whedon
[PRE REVIEW]: Bedparse: feature extraction from BED files
**Submitting author:** @tleonardi (<a href="http://orcid.org/0000-0002-4449-1863">Tommaso Leonardi</a>)
**Repository:** <a href="https://github.com/tleonardi/bedparse" target ="_blank">https://github.com/tleonardi/bedparse</a>
**Version:** v0.2.0
