### 2019-01-23 2125a54
Corrected help message for filter subcommand
