### 2019-01-23 58466c5
Removed python3.7 because not fully supported by Travis
