+++ setup.py
@@ -20,7 +20,6 @@ setup(name='bedparse',
           'Programming Language :: Python :: 3.4',
           'Programming Language :: Python :: 3.5',
           'Programming Language :: Python :: 3.6'
-          'Programming Language :: Python :: 3.7'
       ],
       packages=['bedparse'],
       install_requires=['argparse', 'setuptools'],
