### 2019-01-24 #1196 @tleonardi
Hi @brainstorm - thanks for accepting to be the editor of this paper!
I've noticed that there are a couple of formatting errors in the PDF. Is it ok if I commit the fix and ask whedon to recompile the pdf? Or should I leave it as is till after the reviewers are done?
