### 2019-01-24 08e8074
Paper formatting fixes
