+++ paper.bib
@@ -22,6 +22,5 @@ number = {D1},
 pages = {D754-D761},
 year = {2018},
 doi = {10.1093/nar/gkx1098},
-URL = {http://dx.doi.org/10.1093/nar/gkx1098},
-eprint = {/oup/backfile/content_public/journal/nar/46/d1/10.1093_nar_gkx1098/2/gkx1098.pdf}
+URL = {http://dx.doi.org/10.1093/nar/gkx1098}
 }
