### 2019-01-24 372c8bc
Paper moved to a dedicated branch
