### 2019-01-28 #1196 @tleonardi
Hi @brainstorm,
using @whedon's list I've found a few potential reviewers with expertise in python and genomics:
@dpolychr, @RHagenson, @betteridiot, @jean997, @giraola, @conradstack, @afrubin, @nhejazi
Hopefully someone will be interested in reviewing bedparse!
