### 2019-01-28 a181b2e
Added Sphinx documentation
