+++ docs/index.rst
@@ -1,0 +1,37 @@
+.. bedparse documentation master file, created by
+   sphinx-quickstart on Mon Jan 28 14:46:58 2019.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Bedparse documentation 
+====================================
+Bedparse is a simple python module and CLI tool to perform common operations on BED files.
+
+It offers the following functionality:
+
+* Filtering of transcripts based on annotations
+* Joining of annotation files based on transcript names
+* Conversion from GTF to BED format
+* Conversion from UCSC to Ensembl chromosome names (and viceversa)
+* Conversion from bed12 to bed6
+* Promoter reporting
+* Intron reporting
+* CDS reporting
+* UTR reporting 
+
+
+Contents
+===============
+
+.. toctree::
+   :maxdepth: 2
+
+   Installation
+   Motivation
+   Usage
+
+
+
+
+
+
