+++ docs/Usage.md
@@ -1,0 +1,418 @@
+## Usage
+```text
+> bedparse --help
+usage: bedparse [-h] [--version]
+                {3pUTR,5pUTR,cds,promoter,introns,filter,join,gtf2bed,bed12tobed6,convertChr}
+                ...
+
+Perform various simple operations on BED files.
+
+positional arguments:
+  {3pUTR,5pUTR,cds,promoter,introns,filter,join,gtf2bed,bed12tobed6,convertChr}
+                        sub-command help
+    3pUTR               Prints the 3' of coding genes.
+    5pUTR               Prints the 5' of coding genes.
+    cds                 Prints the CDS of coding genes.
+    promoter            Prints the promoters of transcripts.
+    introns             Prints BED records corresponding to the introns of
+                        each transcript in the original file.
+    filter              Filters a BED file based on an annotation.
+    join                Joins a BED file with an annotation file using the BED
+                        name (col4) as the joining key.
+    gtf2bed             Converts a GTF file to BED12 format.
+    bed12tobed6         Converts a BED12 file to BED6 format
+    convertChr          Convert chromosome names between UCSC and Ensembl
+                        formats
+
+optional arguments:
+  -h, --help            show this help message and exit
+  --version, -v         show program's version number and exit
+```
+
+The basic syntax in the form: `bedparse subcommand [parameters]`.
+
+For a list of all subcommands and a brief explanation of what they do, use: `bedparse --help`
+
+For a detailed explanation of each subcommand and a list of its paramters, use the `--help` options after the subcommand's name, e.g.: `bedparse promoter --help`
+
+---
+
+### 3'/5' UTRs
+
+#### Usage
+```text
+> bedparse 3pUTR --help
+usage: bedparse 3pUTR [-h] [bedfile]
+```
+
+Report the 5' or 3' UTRs of each coding transcript in the BED file. 
+
+UTRs are defined as the region between transcript start/end and CDS start/end (the CDS is in turn defined as the region between thickStart and thickEnd).
+
+Transcripts with an undefined CDS (i.e. with thickStart and thickEnd set to the same value) are not reported.
+
+#### Examples
+```text
+> cat transcripts.bed 
+chr1	167721988	167790819	ENST00000392121.7	0	+	167722151	167787921	0	3	254,167,3000,	0,43594,65831,
+
+> bedparse 3pUTR transcripts.bed 
+chr1	167787921	167790819	ENST00000392121.7	0	+	167787921	167787921	0	1	2898,	0,
+
+```
+
+---
+
+### CDS
+
+#### Usage
+```text
+> bedparse cds --help
+usage: bedparse cds [-h] [--ignoreCDSonly] [bedfile]
+
+Report the CDS of each coding transcript (i.e. transcripts with distinct
+values of thickStart and thickEnd). Transcripts without CDS are not reported.
+
+positional arguments:
+  bedfile          Path to the BED file.
+
+optional arguments:
+  -h, --help       show this help message and exit
+  --ignoreCDSonly  Ignore transcripts that only consist of CDS.
+```
+
+#### Examples
+```text
+> cat transcripts.bed 
+chr1	167721988	167790819	ENST00000392121.7	0	+	167722151	167787921	0	3	254,167,3000,	0,43594,65831,
+
+> bedparse cds transcripts.bed 
+chr1	167722151	167787921	ENST00000392121.7	0	+	167722151	167787921	0	3	91,167,102,	0,43431,65668,
+```
+
+---
+
+### Promoters
[…]
