### 2019-02-02 #1196 @Benjamin-Lee
@pjotrp Happy to review for this
