### 2019-02-06 #1228 @tleonardi
> @tleonardi - I'm not sure why Whedon can't pick up the branch properly. I've compiled the paper locally: [10.21105.joss.01228.pdf](https://github.com/openjournals/joss-reviews/files/2825803/10.21105.joss.01228.pdf)

Thanks @arfon!
