### 2019-02-07 5427a3f
Added -v option to bedparse filter
