### 2019-02-09 #1228 @Benjamin-Lee
As a quick semi-related note, from reading the paper, you might be interested in citing (or signing!) [The Small Tools Manifesto For Bioinformatics](https://github.com/pjotrp/bioinformatics).
