### 2019-02-10 cc69528
Improved error handling in bedline.promoter()
