### 2019-02-10 be74b24
Added validateFormat function, test and docs
