### 2019-02-10 5c4c9fc
Fixed format string to retain compatibility with python<=3.5
