### 2019-02-10 6811f8b
Added initial tutorial as per issue #13
