### 2019-02-10 f6b61b7
self.stranded property converted to boolean
