### 2019-02-11 72deef0
Added diagram, completed tutorial
