+++ docs/tutorial.md
@@ -1,13 +1,15 @@
 # Bedparse tutorial
 
-Hi, thanks for your interest in `bedparse1 `!
+![bedparse functions diagrams](https://github.com/tleonardi/bedparse/blob/devel/docs/bedparse.svg)
+
+Hi, thanks for your interest in `bedparse`!
 
 The following is a short tutorial that will guide you through the functionality of `bedparse`. You can find the `example.bed` file in this repo under docs/example.bed.
-This files contains 6 human transcript models from Gencode. The first three are non-coding transcripts (i.e. field 7 and 8 are the same coordinate), wheread the last three are coding (i.e. fields 7 and 8 indicate the start and end of the [CDS](https://en.wikipedia.org/wiki/Coding_region)).
+This file contains 6 human transcript models from Gencode. The first three are non-coding transcripts (i.e. field 7 and 8 contain the same coordinate), whereas the last three are coding (i.e. fields 7 and 8 indicate the thickStart and thickEnd, i.e. start and end of the [CDS](https://en.wikipedia.org/wiki/Coding_region)).
 
 ## Extracting the promoters
 
-The `bedparse promoter` command reports the promoter of each transcript, defined as user specified interval around the [TSS](https://en.wikipedia.org/wiki/Transcription_start_site).
+The `bedparse promoter` command reports the promoter of each transcript, defined as a user specified interval around the [TSS](https://en.wikipedia.org/wiki/Transcription_start_site).
 For example, we can extract promoters consisting of 1000bp upstream and 500bp downstream of the CDS:
 
 ```
@@ -20,27 +22,39 @@ chr1    924149  925649  ENST00000437963.5
 chr1    924737  926237  ENST00000342066.7
 ```
 
-Note how the TSS (and as a consequence the promoter) depends on the strand: for transcripts on the negative strand the TSS is the end coordinate, i.e. column 3. The `--unstranded` allows you to override this behaviour and report promoters as an interval around column 2, thus disregarding the strand.
+Note how the TSS (and as a consequence the promoter) depends on the strand: for transcripts on the negative strand the TSS is the end coordinate, i.e. column 3. The `--unstranded` option allows you to override this behaviour and report promoters as an interval around column 2, thus disregarding the strand.
 
+## Extracting 5' or 3' UTRs
+
+The [UTRs](https://en.wikipedia.org/wiki/Untranslated_region) are defined in a BED file as the region between the start (column 2) and the thickStart (column 7) for the 5' and between the thickEnd (column 8) and the end (column 3) for the 3'. These rules are reversed for transcripts on the - strand, and `bedparse` automatically takes care of this. Additionally, `bedparse` also handles correctly UTRs that span multiple exons: in these cases `bedparse` recomputes all exon starts and exon lengths as sets coulmns 11 and 12 accordingly.
+
+```
+$ bedparse 5pUTR example.bed 
+chr1    923927  924431  ENST00000420190.6       0       +       923927  923927  0       1       504,    0,
+chr1    925149  925941  ENST00000437963.5       0       +       925149  925149  0       2       40,20,  0,772,
+chr1    925737  925941  ENST00000342066.7       0       +       925737  925737  0       2       63,20,  0,184,
+```
+
+Clearly, as you can see from the output above, UTRs are only reported for coding transcripts.
 
 ## Extracting the CDS
 
 To extract the CDS of the coding transcripts in the BED file use the `bedparse cds` command:
 
 ```
-bedparse cds example.bed 
+$ bedparse cds example.bed 
[…]
