+++ docs/index.rst
@@ -30,7 +30,7 @@ Contents
    Installation
    Motivation
    Usage
-
+   Tutorial
 
 
 
