### 2019-02-11 35fd8d0
Added manifesto
