+++ paper.bib
@@ -24,3 +24,13 @@ year = {2018},
 doi = {10.1093/nar/gkx1098},
 URL = {http://dx.doi.org/10.1093/nar/gkx1098}
 }
+
+@misc{MANIFESTO,
+  author        = {Pjotr Prins},
+  title         = {Small tools MANIFESTO for Bioinformatics},
+  month         = {Feb.},
+  year          = {2014},
+  doi           = "{10.5281/zenodo.11321}",
+  url           = "{http://dx.doi.org/10.5281/zenodo.11321}",
+  howpublished  = {https://github.com/pjotrp/bioinformatics/blob/master/README.md}
+}
