+++ paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 The BED format is a plaintext file format commonly used in bioinformatics to represent genomic features: each line in the file corresponds to a genomic feature (e.g. a gene, transcript, peak, regulatory region, etc.) and consists of up to 12 tab-separated fields that define its genomic coordinates and exon-intron structure. This format is also commonly used to graphically visualise genomic features with genome browser softwares, and is one of the standard formats used by the UCSC Genome Browser [@ucsc2002] and the Ensembl Genome Browser [@ensembl2018]. 
 One of the major advantages of the BED format over many of its alternatives is that each line includes all the information required to define an individual gene/transcript model. This allows to perform most operations on BED files as part of unix pipes, for example using GNU awk one-liners or small custom scripts. This _ad-hoc_ approach, albeit usually simple and effective, often leads to repetition and/or code duplication and can be prone to errors, bugs or typos not always easy to detect.
 
-``Bedparse`` aims to simplify and standardise many of the operations and feature-extractions commonly done on BED files by adhering to the UNIX philosophy of doing one thing and doing it well. Despite the simplicity of many of its functions, bedparse is thouroughly and rigourously tested through an automated test suit to ensure the accuracy and correctness of the results. Additionally, bedparse performs syntax validation checks on the input BED files and warns the user in case of inconsistencies or malformed files.
+``Bedparse`` aims to simplify and standardise many of the operations and feature-extractions commonly done on BED files by adhering to the UNIX philosophy of doing one thing and doing it well. Despite the simplicity of many of its functions, bedparse is thouroughly and rigourously tested through an automated test suit to ensure the accuracy and correctness of the results. Additionally, ``bedparse`` performs syntax validation checks on the input BED files and warns the user in case of inconsistencies or malformed files. ``Bedparse`` embraces the motives of _The small tools manifesto for bioinformatics_[@MANIFESTO] and abides to its rules reccomendations.  
 
 ``Bedparse`` implements the following functions:
 
@@ -40,7 +40,7 @@ In addition to the feature-extraction functions reported above, ``bedparse`` als
 * _gtf2bed_ allows converting Ensembl/Gencode Gene Transfer Format (GTF) files into BED format, with options to specify extra fields to add after column 12.
 * _bed12tobed6_ converts BED12 files to the BED6 format.
 
-Internally, ``bedparse`` implements a _bedline_ class that performs several checks on each BED field in order to ensure the correctness of the format, and implements methods that perform the functions listed above. This type of design allows ``bedparse`` to be either imported by other projects as a python module or to be used as a standalone tool through its command line interface.
+Internally, ``bedparse`` implements a _bedline_ class that performs several checks on each BED field in order to ensure the correctness of the format, and implements methods that perform the functions listed above. This type of design allows ``bedparse`` to be either imported by other projects as a python module or to be used as a standalone tool through its command line interface, either on its own or as part of a _pipe_.
 
 In conclusion, ``bedparse`` is a light, versatile and portable tool developed using good programming practices and a test-driven development approach. Its use as part of bioinformatics pipelines will contribute to speeding up development time and preventing bugs.
 
