### 2019-02-11 #1228 @arfon
> @whedon generate pdf from branch paper

This feature is very buggy I'm afraid and fails for some repositories (looks like yours is one of them).
