### 2019-02-11 #1228 @arfon
Here's another version I compiled locally: [10.21105.joss.01228.pdf](https://github.com/openjournals/joss-reviews/files/2851159/10.21105.joss.01228.pdf)
