### 2019-02-11 #1228 @tleonardi
Hi @brainstorm, I've pushed some edits in response to the comments that @Benjamin-Lee made in the bedparse [repo](https://github.com/tleonardi/bedparse/issues). I've also added the citation to the Manifesto and the community guidelines.
@Benjamin-Lee, do you think I should add a direct statement of need? Any comments on how to best address this point?

@arfon, thanks for compiling locally! Let me know if/how I can help to debug the issue!
