### 2019-02-11 #1228 @tleonardi
@arfon I hadn't pushed the last changes, do you mind compiling the pdf again when you have time? Sorry for the trouble..
