### 2019-02-11 #1228 @arfon
@tleonardi - I fixed the bug :-)
