### 2019-02-11 f6946f7
Some minor edits
