### 2019-02-11 #1228 @Benjamin-Lee
Please note that the current unchecked boxes (functionality and performance) are just due to the fact that I haven't had a chance to review those aspects, not due to any issues I've identified with the software.
