+++ tests/Makefile
@@ -17,9 +17,13 @@ SHELL:=/bin/bash
 all: fivep threep introns gtf2bed operations
 
 .PHONY: data
-data: test_data.tgz
-test_data.tgz: 
-	wget -O test_data.tgz "https://raw.githubusercontent.com/tleonardi/bedparse/devel/tests/test_data.tgz" && tar -zxvf test_data.tgz
+
+test_data.tgz:
+	wget -O test_data.tgz "https://raw.githubusercontent.com/tleonardi/bedparse/devel/tests/test_data.tgz"
+
+data: ensembl.bed 
+ensembl.bed: test_data.tgz
+	tar -zxvf test_data.tgz
 
 .PHONY: fivep
 fivep: data
