### 2019-02-18 1ff426f
Added licence badge
