### 2019-02-18 9afbe6e
Updated Travis badge to track devel branch
