### 2019-02-19 b43a17c
Implemented stranded option to tx2genome
