### 2019-02-21 c968072
Fixed bug in tx2genome() stranded handling
