### 2019-02-23 #1228 @tleonardi
Hi @brainstorm, I have added a small feature to bedparse and done a couple of changes following @Benjamin-Lee's feedback. I'd like to make a new release because I have other project's depending on bedparse that require this new feature. Since there's a "version" checkbox, I was wondering whether it's ok for me to make a release at this stage...
