### 2019-02-25 #1228 @brainstorm
Yes it should be ok as long as @Benjamin-Lee reviews the changes accordingly.
