+++ setup.py
@@ -7,7 +7,7 @@ setup(name='bedparse',
       description='A simple library and CLI tool to manipulate BED files',
       long_description=long_description,
       long_description_content_type="text/markdown",
-      version="0.2.0",
+      version="0.2.1",
       url='https://github.com/tleonardi/bedparse',
       author='Tommaso Leonardi',
       author_email='tom@tleo.io',
