### 2019-02-25 #1228 @tleonardi
Great, thanks. I've released v0.2.1.
Actually, I've just now noticed that @Benjamin-Lee ticked all the boxes above. If he's happy with the changes this could become the final release once he is done with the review?
