### 2019-02-25 7a36eed
Typo fixes and minor grammar edits
Please also note that there is a missing noun in Line 24. Because the choice of noun can affect the meaning of the sentence, I am leaving that choice up to you. All of the other changes are for stylistic consistency, grammatical correctness, readability or concision.
