### 2019-02-25 #1228 @Benjamin-Lee
I just looked over the changes and they look good.

I also just noticed that you say `Bedparse is a simple python module and CLI tool` in the paper and say it can be imported but there aren't any docs about that. Since it can be imported, there need to be API-level docs.

In addition, I have a few copy edits, which I submitted as a [PR](https://github.com/tleonardi/bedparse/pull/16).
