### 2019-02-25 5f7a250
Improved docstring for tx2genome
