### 2019-02-25 b3416a1
Improved docstrings
