### 2019-02-25 db433b1
Added API documentation
