### 2019-02-25 #1228 @tleonardi
Thanks @Benjamin-Lee for pointing out the missing docs. I've written the documentation as docstrings and compiled it with sphinx autodoc. It has it's own [page](https://bedparse.readthedocs.io/en/devel/bedparse.bedline.html) on Read the Docs and simple [example](https://bedparse.readthedocs.io/en/devel/Tutorial.html#apis) in the tutorial.
I've also merged the paper's PR.
Since I've edited the code to complete the docstrings I'll make a new release when the review process is completed.
