### 2019-02-25 #1228 @Benjamin-Lee
So sorry, I was looking at the master branch and didn't notice that the devel one has the docs. Is it possible to get method-level examples in the API docs using `:Example:`? (As an aside, you might be interested in [`napoleon`](http://www.sphinx-doc.org/en/master/usage/extensions/napoleon.html). which makes writing sphinx docstrings more pleasant)

Also, what is the reasoning behind using `pprint` and `print` methods over the `__str__()` and `__repr__()` methods and the built-in print function? There seems to be some code duplication between `__str__()` and `__print()` but it may be that I am missing something obvious (like being on the right branch).

The code is well commented and tested. You might be interested in using an automatic formatter such as [black](https://github.com/ambv/black) or [autopep8](https://github.com/hhatto/autopep8) to keep the style consistent and a tool such as [CodeFactor](https://www.codefactor.io) or [Codacy](https://www.codacy.com) for automated code review.

One minor issue is that `translateChr` shadows the built-in `all` keyword, which [can lead to bad things happening](https://docs.quantifiedcode.com/python-anti-patterns/correctness/assigning_to_builtin.html).

These are small quibbles, however. This tool is useful, documented, tested, and tested. @tleonardi, thank you for your contribution to the bioinformatics community!

@brainstorm and @arfon, the paper is ready to go. In the proof, the URL for the small tools manifesto overruns the edge of the page (see below). Also, the use of the month in one of the citations but not the others is a little bit jarring. These seem to be template-related issues.
