+++ paper.bib
@@ -28,9 +28,7 @@ URL = {http://dx.doi.org/10.1093/nar/gkx1098}
 @misc{MANIFESTO,
   author        = {Pjotr Prins},
   title         = {The Small Tools Manifesto for Bioinformatics},
-  month         = {Feb.},
   year          = {2014},
   doi           = {10.5281/zenodo.11321},
-  url           = {http://dx.doi.org/10.5281/zenodo.11321},
-  howpublished  = {https://github.com/pjotrp/bioinformatics/blob/master/README.md}
-}
+  url           = {http://dx.doi.org/10.5281/zenodo.11321}
+ }
