### 2019-02-26 #1228 @arfon
@Benjamin-Lee @tleonardi this PR should fix the reference formatting: https://github.com/tleonardi/bedparse/pull/17
