### 2019-02-26 tleonardi/bedparse#18 @tleonardi
Get rid of print() and pprint()
The print() and pprint() methods should be removed and their functionality implemented in __str__ and __repr__.
See thread in JOSS [review](https://github.com/openjournals/joss-reviews/issues/1228#issuecomment-467196907)
