### 2019-02-26 a255138
Fixed shadowing of built-in `all`
