### 2019-02-26 #1228 @tleonardi
No need to be sorry @Benjamin-Lee, I only wrote the docs after you pointed out they were missing! :)
Thanks for the careful review and the new suggestions. I have just renamed the `all` argument in `translateChr()`, thanks for spotting it!
Regarding the  `pprint` and `print` methods, you are right, I should get rid of them. A while back they made sense, but I have been meaning to replace them with  `__str__()` and `__repr__()` for a while... the only complication is that I have quite a few pieces of code in other projects that use those functions, so I have just been keeping them around for backward compatibility. I have opened an [issue](https://github.com/tleonardi/bedparse/issues/18#issue-414539816), I'll refactor as soon as I get a chance.
As for the examples in the docstrings, I'm working on it now.

Thanks @arfon for the PR! It's now merged.
