### 2019-02-26 3766b0c
Added API examples in docstrings
