+++ bedparse/bedline.py
@@ -123,14 +123,18 @@ class bedline(object):
     def promoter(self, up=500, down=500, strand=True):
         """ Returns the promoter of a bedline object
 
-        :param up: Number of upstream bases
-        :param down: Number of donwstream bases
-        :param strand: If false strandedness is ignored 
-        :type up: int
-        :type down: int
-        :type strand: bool
-        :returns: bedline
+        Args:
+            up (int): Number of upstream bases
+            down (int): Number of donwstream bases
+            strand (bool): If false strandedness is ignored 
+        Returns:
+            bedline: The promoter as a bedline object
+        Examples:
+	    >>> bl = bedline(['chr1', 1000, 2000, 'Tx1', '0', '+'])
+            >>> print(bl.promoter())
+            ['chr1', 500, 1500, 'Tx1']
         """
+
         if strand and self.bedType<6:
             raise BEDexception("You requested stranded promoters, but the BED file appears to be unstranded")
         if not strand or self.strand=="+":
@@ -146,9 +150,15 @@ class bedline(object):
     def utr(self, which=None):
         """ Returns the UTR of coding transcripts (i.e. those with a CDS) 
         
-        :param which: Which UTR to return: 3 for 3'UTR or 5 for 5' UTR
-        :type which: int
-        :returns: bedline
+	Args:
+	    which (int): Which UTR to return: 3 for 3'UTR or 5 for 5' UTR
+        Returns:
+	    bedline: The UTR as a bedline object
+	Examples:
+	    >>> bl = bedline(["chr1", 100, 500, "Tx1", 0, "+", 200, 300, ".", 1, "400,", "0,"])
+            >>> print(bl.utr(which=5))
+            ['chr1', 100, 200, 'Tx1', 0, '+', 100, 100, '.', 1, '100,', '0,']
+
         """
         if(not self.stranded):
             raise BEDexception("UTRs for an unstranded transcript make little sense: "+self.name)
@@ -247,10 +257,15 @@ class bedline(object):
 
     def cds(self, ignoreCDSonly=False):
         """Return the CDS of a coding transcript. Transcripts without CDS are not reported
-
-        :param ignoreCDSonly: If True return None when the entire transcript is CDS 
-        :type ignoreCDSonly: bool
-        :returns: bedline
+	
+	Args:
+            ignoreCDSonly (bool): If True return None when the entire transcript is CDS 
+	Returns:
+	    bedline: The CDS as a bedline object
+	Examples:
+	    >>> bl = bedline(["chr1", 100, 500, "Tx1", 0, "+", 200, 300, ".", 1, "400,", "0,"])
+            >>> print(bl.cds())
+            ['chr1', 200, 300, 'Tx1', 0, '+', 200, 300, '.', 1, '100,', '0,']
         """
         if(not self.stranded):
             raise BEDexception("CDS for an unstranded transcript makes little sense: "+self.name)
@@ -310,7 +325,15 @@ class bedline(object):
     def introns(self):
         """ Returns a bedline object of the introns of a transcript
         
-        :returns: bedline
+        Returns:
+	    bedline: The introns of the transcripts as a bedline object
+	Examples:
+	    >>> bl = bedline(["chr1", 100, 420, "Name", 0, "+", 210, 310, ".", 4, "20,20,20,20,", "0,100,200,300,"])
+            >>> print(bl.introns())
+            ['chr1', 120, 400, 'Name', 0, '+', 120, 120, '.', 3, '80,80,80,', '0,100,200,']
+            >>> bl = bedline(["chr1", 100, 420, "Name", 0, "-", 210, 310, ".", 1, "320,", "0,"])
+            >>> print(bl.introns())
+            None
         """
[…]
