+++ docs/conf.py
@@ -40,7 +40,7 @@ release = 'v0.2.0'
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.autodoc']
+extensions = ['sphinx.ext.napoleon']
 autoclass_content = 'both'
 
 # Add any paths that contain templates here, relative to this directory.
