### 2019-02-26 #1228 @tleonardi
Ok, I have added the examples for the API methods in the docstrings. I've also switched to [`napoleon`](http://www.sphinx-doc.org/en/master/usage/extensions/napoleon.html). Thanks @Benjamin-Lee for the tip, the syntax is much more readble!
