### 2019-02-26 478099c
Added changes in 0.2.2
