### 2019-02-26 caeba02
Consistent use of hyphenation and a few nitpicks
See [https://www.grammarbook.com/punctuation/hyphens.asp](https://www.grammarbook.com/punctuation/hyphens.asp)
