### 2019-02-26 #1228 @tleonardi
Ok, this last one is the version incorporating all the changes suggested by @Benjamin-Lee.
@brainstorm, is there anything else required on my side? I'm ready to make a new release with the last batch of changes.
