### 2019-02-27 #1228 @brainstorm
Looks good to me, feel free to do that release, pack it and return with a zenodo doi for archival.
