### 2019-02-27 58374d7
Added basic usage and badges
