### 2019-02-27 #1228 @tleonardi
Ok, I've just release v0.2.2.
DOI: [10.5281/zenodo.2578820](https://doi.org/10.5281/zenodo.2578820)
thanks!
