### 2019-02-27 c5e047a
Added Zenodo DOI badge
