### 2019-02-27 da93b3d
Added Zenodo DOI badge
