+++ README.md
@@ -1,6 +1,7 @@
 [![Build Status](https://travis-ci.org/tleonardi/bedparse.svg?branch=master)](https://travis-ci.org/tleonardi/bedparse)
-[![Docs Status](https://readthedocs.org/projects/bedparse/badge/?version=master&style=flat)](https://bedparse.readthedocs.io/en/master/)
+[![Docs Status](https://readthedocs.org/projects/bedparse/badge/?version=latest&style=flat)](https://bedparse.readthedocs.io/en/latest/)
 [![JOSS Status](http://joss.theoj.org/papers/22763a3b37fde13e548e884edd3221fa/status.svg)](http://joss.theoj.org/papers/22763a3b37fde13e548e884edd3221fa)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.2578820.svg)](https://doi.org/10.5281/zenodo.2578820)
 [![License: MIT](https://img.shields.io/badge/License-MIT-brightgreen.svg)](https://opensource.org/licenses/MIT)
 
 # Bedparse
