### 2019-02-27 #1228 @arfon
> PDF failed to compile for issue #1228 with the following error:

@tleonardi @brainstorm - the paper must be in the `master` branch for us to accept the submission.
