### 2019-02-27 5640b8a
Added JOSS paper
