### 2019-02-27 #1228 @tleonardi
@arfon It looks like whedon is not picking up the second reference (Prins 2014). Is this normal?
