### 2019-02-27 #1228 @arfon
> @arfon It looks like whedon is not picking up the second reference (Prins 2014). Is this normal?

Yeah, I think we only check article DOIs at this point.
